% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-section.R
\name{crunchbase_expand_section}
\alias{crunchbase_expand_section}
\title{Expand a section of a node detail}
\usage{
crunchbase_expand_section(node, relationship, ...)
}
\arguments{
\item{node}{a downloaded node, or multiple nodes as returned by 
\code{crunchbase_get_details}}

\item{relationship}{character: the name(s) of the section(s) to expand - 
vectors with length longer than 1 are treated as multiple sections to expand}

\item{...}{other arguments passed to crunchbase_get_collection}
}
\description{
Node details include numerous sections. For instance, an organization record 
may include a "current team" section. These sections will only include the 
first 8 records in the section, along with a path to the full collection. 
This function calls \code{crunchbase_get_collection} to pull the entire 
collection.
}
\examples{
crunchbase_get_details("organization/facebook") \%>\%
    crunchbase_expand_section("current_team")
}

