% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crunchbase_flatten_collection}
\alias{crunchbase_flatten_collection}
\title{Flatten a collections object}
\usage{
crunchbase_flatten_collection(collection)
}
\arguments{
\item{collection}{a downloaded collections object, e.g. as returned
by crunchbase_get_collection}
}
\description{
Each page of a collections object is returned as an element of a list. This
function flattens it out into one long data frame. Note that if you use
\code{flatten=TRUE} in \code{crunchbase_get_collection}, then you don't need
to use this function
}
\examples{
## a list of all organizations listed in the CrunchBase database
crunchbase_get_collection("organizations") \%>\% crunchbase_flatten_collection
}

