% Generated by roxygen2 (4.0.1): do not edit by hand
\name{crunchbase_GET}
\alias{crunchbase_GET}
\title{Query CrunchBase.}
\usage{
crunchbase_GET(path, ...)
}
\arguments{
\item{path}{A query endpoint, presented either as a single string
 (e.g. "people" or "person/johndoe") or as a vector that drills down a
 hierarchy (e.g. c("person", "johndoe"))}

\item{...}{any other query parameters, each entered as parameter = "value"}
}
\description{
Query CrunchBase via a GET request. All of the CrunchBase API's
functionality comes through GET requests. This function collects any
query parameters and GETs the resulting URL
}
\examples{
x <- crunchbase_GET(c("person", "bill-gates"))
x <- crunchbase_GET("person/bill-gates")
crunchbase_GET("organizations", location="San Francisco")
}

