% Generated by roxygen2 (4.0.1): do not edit by hand
\name{crunchbase_get_collection}
\alias{crunchbase_get_collection}
\title{Get a collections endpoint}
\usage{
crunchbase_get_collection(path, flatten = TRUE, page.limit = FALSE,
  delay = 1.3, ...)
}
\arguments{
\item{path}{the path for the collections endpoint}

\item{page.limit}{use if you only want a subset of the pages}

\item{delay}{length of time (in seconds) to pause between each GET request,
useful to avoid going over usage limits.}

\item{...}{other arguments passed to crunchbase_GET}
}
\description{
Simplifies GETting multi-page collections endpoints. Instead of making
repeated calls to \code{crunchbase_GET} with \code{page=n}, just use this.
}
\examples{
crunchbase_get_collection(c("organization","facebook","current_team"))
crunchbase_get_collection("people") ## returns all people in CrunchBase
crunchbase_get_collection("organizations", location="Oakland")
}

