% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crunchbase_get_entities}
\alias{crunchbase_get_entities}
\title{Get one or more entity endpoints}
\usage{
crunchbase_get_entities(paths, exclude = no_exclude, delay = 1.3, ...)
}
\arguments{
\item{paths}{a vector or list of paths, for instance
c("person/some-name", "person/another-name", "person/somebody") or
list(c("person", "some-name"), c("person", "another-name"),
c("person", "somebody")). This can also be a data.frame with a column named
"path," (for instance: the result of crunchbase_get_collection)}

\item{exclude}{any function that takes as its input a crunchbase entity
and returns as its output either \code{TRUE} or \code{FALSE}}

\item{delay}{length of time in seconds to wait between GET requests, useful
to avoid hitting usage limits}

\item{...}{other arguments passed to \code{crunchbase_GET}}
}
\description{
Retrieve data about one or more CrunchBase entities. This is a convenient
wrapper around the \code{crunchbase_GET} function, with the ability to
wait between requests (to avoid usage limits) and exclude entities from
your results.
}
\details{
The return value will be a list where each element is an entity detail
node. \code{crunchbase_get_entities} is useful in conjunction with the
results of \code{crunchbase_get_collection} (see examples)
}
\examples{
## see details on all companies in Oakland
crunchbase_get_collection("organizations", location="Oakland") \%>\%
    crunchbase_get_entities
}

