% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crunchbase_key}
\alias{crunchbase_key}
\title{Set or get a CrunchBase user key.}
\usage{
crunchbase_key(force = FALSE)
}
\arguments{
\item{force}{Overwrite existing key? If TRUE, the user will be prompted to
enter a user key even if one has already been entered before.}
}
\description{
Reads (or writes) the user key to an environment variable.
}
\details{
Request a user key from CrunchBase \url{https://developer.crunchbase.com/}
in order to use the API. This key gets appended to GET queries via the
\code{user_key} parameter. The key will be saved so you don't have to enter
it repeatedly.
}

