% Generated by roxygen2 (4.0.1): do not edit by hand
\name{crunchbase_order}
\alias{crunchbase_order}
\title{Building order parameters for collections queries.}
\usage{
crunchbase_order(mod = "created_at", type = "asc")
}
\arguments{
\item{mod}{Ordering function (eg created_at, modified_at, etc)}

\item{type}{Order (asc or desc)}
}
\description{
This is a pretty useless helper function.
}

