% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_actions.R
\name{crunchbase_parse}
\alias{crunchbase_parse}
\title{Parse raw CrunchBase API responses.}
\usage{
crunchbase_parse(req)
}
\description{
You only need to use this if you're using the basic /code{crunchbase_GET} function.
Other functions automatically parse the responses.
}
\examples{
x <- crunchbase_GET(c("person", "bill-gates"))
crunchbase_parse(x)
}

