% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extracts.R
\name{crunchbase_strip}
\alias{crunchbase_strip}
\title{Collapse a crunchbase node into a spreadsheet}
\usage{
crunchbase_strip(node)
}
\arguments{
\item{node}{the node object to be collapsed}
}
\description{
Crunchbase nodes are represented as hierarchical lists. Sometimes, you just
want to see a summarized table version of the data. This function collapses
a single node into a data.frame
}
\examples{
crunchbase_GET("person/dan-rubinstein") \%>\% crunchbase_parse \%>\% crunchbase_strip
}

