% Generated by roxygen2 (4.0.1): do not edit by hand
\name{crunchbase_strip_list}
\alias{crunchbase_strip_list}
\title{Collapse a list of nodes into one data.frame}
\usage{
crunchbase_strip_list(nodes)
}
\description{
Uses \code{crunchbase_strip} on every element of a list of nodes, and then
returns the result as one \code{data.frame}, which can be exported or whatever.
}

