% Generated by roxygen2 (4.0.2): do not edit by hand
\name{no_filter}
\alias{no_filter}
\title{Don't filter anything}
\usage{
no_filter(node)
}
\description{
This is the default exclusion function for \code{crunchbase_get_entities},
there isn't any reason to call it otherwise. It always returns \code{FALSE},
meaning don't exclude anything.
}

