% Generated by roxygen2 (4.0.2): do not edit by hand
\name{word_filter}
\alias{word_filter}
\title{Construct filter functions based on a word or pattern.}
\usage{
word_filter(section, pattern, ignore.case = TRUE)
}
\arguments{
\item{section}{which section of the node is used for the exclusion?}

\item{pattern}{a string or regular expression to look for within the section}

\item{case.insensitive}{TRUE/FALSE, defaults to TRUE, meaning the search is
case-insensitive.}
}
\description{
Construct filter functions based on a word or pattern.
}

