% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crunchbase_expand_section}
\alias{crunchbase_expand_section}
\title{Expand a section of an entity detail}
\usage{
crunchbase_expand_section(node, relationship, ...)
}
\arguments{
\item{node}{a downloaded entity node}

\item{relationship}{character: the name of the section to expand}

\item{...}{other arguments passed to crunchbase_get_collection}
}
\description{
Node details include numerous sections. For instance, an organization record
may include a "current team" section. These sections will only include the
first 8 records in the section, along with a path to the full collection.
This function calls \code{crunchbase_get_collection} to pull the entire
collection.
}
\examples{
crunchbase_get_entities(c("organization", "facebook")) \%>\%
    crunchbase_expand_section("current_team")
}

