% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic_actions.R
\name{crunchbase_GET}
\alias{crunchbase_GET}
\title{Query CrunchBase.}
\usage{
crunchbase_GET(path, ...)
}
\arguments{
\item{path}{A query endpoint, presented either as a single string
 (e.g. "people" or "person/johndoe") or as a vector that drills down a
 hierarchy (e.g. c("person", "johndoe"))}

\item{...}{any other query parameters, see
\url{https://developer.crunchbase.com/docs} for details}
}
\description{
Query CrunchBase via a GET request. All of the CrunchBase API's
functionality comes through GET requests. This function collects any
query parameters and GETs the resulting URL
}
\note{
\code{crunchbase_GET} keeps track of API calls and will automatically
pause a query for the appropriate amount of time in order to keep from making
more than 44 calls per minute. Furthermore, results are cached for the length
of your R session (using \code{memoise}). To make it easier to use the API
collections filters (currently only applies to organization queries), you can
pass \code{data.frame}s directly to \code{crunchbase_GET}, as long as it
contains columns named "type" and "uuid," as the \code{locations} and
\code{categories} collections do.
}
\examples{
x <- crunchbase_GET(c("person", "bill-gates"))
x <- crunchbase_GET("person/bill-gates")
}

