% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crunchbase_get_collection}
\alias{crunchbase_get_collection}
\title{Get a collections endpoint}
\usage{
crunchbase_get_collection(path, speed = 44L, ...)
}
\arguments{
\item{path}{the path for the collections endpoint}

\item{speed}{Number of API calls to make per minute. Default 44.}

\item{...}{other arguments passed to crunchbase_GET}
}
\description{
Simplifies GETting multi-page collections endpoints. Instead of making
repeated calls to \code{crunchbase_GET} with \code{page=n}, just use this.
}
\examples{
crunchbase_get_collection(c("organization","facebook","current_team"))
crunchbase_get_collection("people") ## returns all people in CrunchBase
crunchbase_get_collection("organizations", location="Oakland")
}

