% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-collection.R
\name{crunchbase_get_collection}
\alias{crunchbase_get_collection}
\title{Get a collections endpoint}
\usage{
crunchbase_get_collection(path, ...)
}
\arguments{
\item{path}{the path for the collections endpoint.}

\item{...}{other arguments passed to crunchbase_GET}
}
\description{
Simplifies GETting multi-page collections endpoints. Instead of making 
repeated calls to \code{crunchbase_GET} with \code{page=n}, just use this.
}
\examples{
crunchbase_get_collection(c("organization","facebook","current_team"))
crunchbase_get_collection("people") ## returns all people in CrunchBase

}

