% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-details.R
\name{crunchbase_get_details}
\alias{crunchbase_get_details}
\title{Get one or more entity endpoints}
\usage{
crunchbase_get_details(paths, df_path = NULL, filter = no_filter, ...)
}
\arguments{
\item{paths}{a vector or list of paths, for instance
c("person/some-name", "person/another-name", "person/somebody") or
list(c("person", "some-name"), c("person", "another-name"),
c("person", "somebody")). This can also be a data.frame (for instance: the
result of crunchbase_get_collection)}

\item{df_path}{character if \code{paths} is a data.frame, this should be the
name or integer index of the column containing the query paths. Defaults to
"path"}

\item{filter}{any function that takes as its input a crunchbase node detail
and returns as its output either \code{TRUE} (include) or \code{FALSE}}

\item{...}{other arguments passed to \code{crunchbase_GET}}
}
\description{
Retrieve data about one or more CrunchBase entities. This is a convenient
wrapper around the \code{crunchbase_GET} function, with the ability to
wait between requests (to avoid usage limits) and exclude entities from
your results.
}
\details{
The return value will be a list where each element is an entity detail
node. \code{crunchbase_get_details} is useful in conjunction with the
results of \code{crunchbase_get_collection} (see examples)
}

